+++ src/protocolsupport/protocol/packet/middle/clientbound/play/MiddleBlockChangeSingle.java
@@ -3,8 +3,8 @@ package protocolsupport.protocol.packet.middle.clientbound.play;
 import io.netty.buffer.ByteBuf;
 import protocolsupport.protocol.ConnectionImpl;
 import protocolsupport.protocol.serializer.VarNumberSerializer;
-import protocolsupport.protocol.storage.netcache.ChunkCache;
-import protocolsupport.protocol.storage.netcache.ChunkCache.CachedChunk;
+import protocolsupport.protocol.storage.netcache.chunk.CachedChunk;
+import protocolsupport.protocol.storage.netcache.chunk.ChunkCache;
 import protocolsupport.protocol.types.ChunkCoord;
 
 public abstract class MiddleBlockChangeSingle extends MiddleBlock {
@@ -28,12 +28,12 @@ public abstract class MiddleBlockChangeSingle extends MiddleBlock {
 		int x = position.getX();
 		int y = position.getY();
 		int z = position.getZ();
-		if (y >= 256) { // filter invalid block positions
+		if (y >= 256) {
 			return false;
 		}
 		CachedChunk cachedChunk = cache.getChunkCache().get(ChunkCoord.fromGlobal(x, z));
 		if (cachedChunk != null) {
-			cachedChunk.setBlock(y >> 4, CachedChunk.getBlockIndex(x & 0xF, y & 0xF, z & 0xF), id);
+			cachedChunk.setBlock(y >> 4, CachedChunk.getBlockIndex(x & 0xF, y & 0xF, z & 0xF), (short) id);
 			return true;
 		} else {
 			return false;
